Summarize possible test scenarios.
